# server/api/v1/system/sys_initdb.go
package system

import (
	"fmt"
	"github.com/lzsgo/lzs-go-admin/server/global"
	"github.com/lzsgo/lzs-go-admin/server/model/common/response"
	"github.com/lzsgo/lzs-go-admin/server/model/system/request"
	"go.uber.org/zap"

	"github.com/gin-gonic/gin"
)

type DBApi struct{}

// InitDB
// @Tags InitDB
// @Summary 初始化用户数据库
// @Produce  application/json
// @Param data body request.InitDB true "初始化数据库参数"
// @Success 200 {object} response.Response{data=string} "初始化用户数据库"
// @Router /init/initdb [post]
func (i *DBApi) InitDB(c *gin.Context) {
	if global.GVA_DB != nil {
		global.GVA_LOG.Error("已存在数据库配置!")
		response.FailWithMessage("已存在数据库配置", c)
		return
	}
	var dbInfo request.InitDB
	if err := c.ShouldBindJSON(&dbInfo); err != nil {
		global.GVA_LOG.Error("参数校验不通过!", zap.Error(err))
		response.FailWithMessage("参数校验不通过", c)
		return
	}
	if err := initDBService.InitDB(dbInfo); err != nil {
		global.GVA_LOG.Error("自动创建数据库失败!", zap.Error(err))
		response.FailWithMessage("自动创建数据库失败，请查看后台日志，检查后在进行初始化", c)
		return
	}
	response.OkWithMessage("自动创建数据库成功", c)
}

// CheckDB
// @Tags CheckDB
// @Summary 初始化用户数据库
// @Produce  application/json
// @Success 200 {object} response.Response{data=map[string]interface{},msg=string} "初始化用户数据库"
// @Router /init/checkdb [post]
func (i *DBApi) CheckDB(c *gin.Context) {
	var (
		message  = "前往初始化数据库"
		needInit = true
	)

	if global.GVA_DB != nil {
		message = "数据库无需初始化"
		needInit = false
	}
	global.GVA_LOG.Info(message)
	response.OkWithDetailed(gin.H{"needInit": needInit}, message, c)
}

// ResetDB
// @Tags ResetDB
// @Summary 重置数据库
// @Produce  application/json
// @Success 200 {object} response.Response{data=map[string]interface{},msg=string} "初始化用户数据库"
// @Router /init/resetDB [post]
func (i *DBApi) ResetDB(c *gin.Context) {

	if global.GVA_DB == nil {
		response.FailWithMessage("请先初始化数据库", c)
		return
	} else {
		deleteDatabaseSql := fmt.Sprintf("DROP DATABASE `%s`", global.GVA_CONFIG.Mysql.Dbname)
		global.GVA_DB.Exec(deleteDatabaseSql)
		dbInfo := request.InitDB{
			DBType:   "mysql",
			Host:     global.GVA_CONFIG.Mysql.Path,
			Port:     global.GVA_CONFIG.Mysql.Port,
			UserName: global.GVA_CONFIG.Mysql.Username,
			Password: global.GVA_CONFIG.Mysql.Password,
			DBName:   global.GVA_CONFIG.Mysql.Dbname,
		}
		if err := initDBService.InitDB(dbInfo); err != nil {
			global.GVA_LOG.Error("自动创建数据库失败!", zap.Error(err))
			response.FailWithMessage("自动创建数据库失败，请查看后台日志，检查后在进行初始化", c)
			return
		}
		response.OkWithMessage("重置成功", c)
		return
	}

}