# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Server pets",
        "title": "Swagger Example API",
        "contact": {},
        "version": "0.0.1"
    },
    "paths": {
...
        "/init/initdb": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "InitDB"
                ],
                "summary": "初始化用户数据库",
                "parameters": [
                    {
                        "description": "初始化数据库参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.InitDB"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "初始化用户数据库",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.InitDB": {
            "type": "object",
            "required": [
                "dbName",
                "userName"
            ],
            "properties": {
                "dbName": {
                    "description": "数据库名",
                    "type": "string"
                },
                "dbType": {
                    "description": "数据库类型",
                    "type": "string"
                },
                "host": {
                    "description": "服务器地址",
                    "type": "string"
                },
                "password": {
                    "description": "数据库密码",
                    "type": "string"
                },
                "port": {
                    "description": "数据库连接端口",
                    "type": "string"
                },
                "userName": {
                    "description": "数据库用户名",
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}